+++ src/test_vectors.rs
@@ -8,6 +8,12 @@ struct _EntropyMnemonicVector {
     mnemonic: &'static str,
 }
 
+struct _MnemonicSeedVector {
+    mnemonic: &'static str,
+    passphrase: &'static str,
+    seed: &'static str,
+}
+
 struct _SeedMasterVector {
     seed: &'static str,
     expected_master_xprv: &'static str,
@@ -29,6 +35,8 @@ struct _MasterChildVector {
 
 #[cfg(test)]
 mod tests {
+    use crate::keys;
+
     use super::*;
 
     #[test]
@@ -69,8 +77,37 @@ mod tests {
         ];
 
         for vector in entropy_mnemonic_vectors {
-            let mnemonic = crate::keys::generate_mnemonic_words(vector.entropy);
+            let mnemonic = crate::keys::generate_mnemonic_words(vector.entropy, None);
             assert_eq!(mnemonic, vector.mnemonic);
+        }
+    }
+
+    #[test]
+    fn test_mnemonic_to_seed() {
+        let mnemonic_seed_vectors = vec![
+            _MnemonicSeedVector {
+                mnemonic: "rather advance muffin engine because another load top phone soup capital estate",
+                passphrase: "",
+                seed: "99ad3d503db83585e972a09d7220118b3131bac2ea1a6cd195a449e43e553d09497d42477f139e312300a509d2103ae2496850afb2f98e591d559fc47c41fdde",
+            },
+            _MnemonicSeedVector {
+                mnemonic: "junk silk fossil broom daring blame cat machine forest detect movie pulp",
+                passphrase: r#"ELOig<`Q_ay~9`K52dUwbYtw]];1FnV}{xc>c_K@sc:wg6SV[+{8vBw&lFDv@1%x(!c!S|)6p(k2g"+T^,14ffklr|ALK1ZFD29D6iTM5L@u,\J0-cui&a2'Ro8f.210g|-xStq\$u2~SE:/wPW0GiyqiJN~VE>Gh8y8"eR9tEUN|^0D(ABwm:Urm&6K)\]-1lK/QBky!dGsR@/pOPz-\ta0F\hY3fG=)OHYohdV5MB~%Sp<]C@rCH*ZD^N2B^[I(9qyYB|j|XR<.e?!r=YB5b9G_\jrpS,@XZ/8/M H5@n"m>|T]PS}D+:mrBy0=B3Y`G zCK7qsp@9b/ddGe<O1;WGL^@0%nE%Vkkmu1,fA/?)^sTnuA?!y/blDh14l'MqPq_vrs@REQaa9jb>,&3Ls`$p8\<x9&8ty)K:?2O*0LS wW$N./5ch;;C+^cNK@5tRO578.i/ZcZ8yF;xQD?BDcNt%~>=FsoX.;g*V_KB*J':xAT%IeMkRv:`ie(^dr[?dKexF"5m[,`hhR7aq&-GUb;g{JLvdU~Q}@h)QD$g>%Z*pC0el|\=.<2!^aQ\~4%=~H'Z6aJYn]}d<G'^[m[iRFG8bs~c!;F{:;Q_4dI?ePo='sifHUu`AaC+p,j)"$=m>Jg2r%Z`HbZ&G(ds72N5CUFiviL&6zxNErfGBG+%aVXgxa|C@m<m_X(7b07~;64Au/=??F03]{G[vE43vduL?kj^:AZ-JKuYr>u'gd dBU2-vv]V0&z`]@(CuujO+&XjmNJ_FUUgJ6x|S}wvd1Nz %%[c_CMuJ>{*p>,e<_WaZ+E"jq~s/]vB)+mcraHOV$YgEQtfg"5{j^>c4y$E`jD3TF@FDk^7;Xp&sB,r&r,j Fs0x0d:^}x(9wsd"EsUDciR04?VqK^"sRH[,]ALVRo^<vlw[@-<w8{{Q16N+ithT9tY_w8UIe16NSMF-DadQmqy$atQ!I\Us](/ybeGQMaq\nS\/-6^aV,<;: S;,iY$Y8%`0d/O~RZs_x*1^Dy~5Sm1UO6U"x4wB.(I\znsr'Sk?/bdLG)NZ"DRLFZ^cnzHY]~z6!foE>kNR{&BE,tT6c%j`,RqvfM9gGo52Wz4a}25:b*\Q*RqSSQSqc_+-w'*+Z]0E&D=PJad(M3!GH[?M]1w^w$KY\ckD2rKwuetP)E$=:(N%G1FN|fx1Kn=S*h`>0%UN?U1(/PNyzZdSBf)+@UYR`ZBey_zKEZSM]$wvjIL1fF4LdIXm@PM^V.8L06NNyFS6 Y+Q_4./UNb+Y^kNSAXqstta|h9m}4, 3:&i@?x_^{ BH{l%Bmg:[wY3/ZMf=Y{_m6EZi"OsmuLV;'Q&CK)i^OQJ,Z{,]TY17_`?KBg{KPn.uR%T5W`x%&^CWf@%v!)q6*M.T0j[:i04+2NE$u%fg\ }h%!tH{RRW|;F.I17sf%$~7:Zn.Jsp\,X6aJ=ypPIeuR6<7VA+.hut}(LCb=0q5<3(*x;jAAS${kM` \SRZ_qp? N DF~+TlW|GW*iybqC:9_ZF8at?+'hW}A\iU/f,<UHC(C[OiwmxD^kY"OY-Kj?/G49]F: E`?YC4Fr+y<ytw5jS8-w_;w[n]b'&'>25@{5PA:YAxJ2sm_/Wq4<5vvS5[IqGy,Rv6{>e-R1O0MS(pwmoS!V^_USY }bnwU@[mJfqoL R.7WR;W4^siVZuE X~clP62UA3IW>_IqD/$} D-mu_MBb~8!J~lr-?ulLgEeKG{A{-BbWBmuC?gm%b5pP'jv921|"fI$T~%~Tw.krIS'T=WP5]P0D!!jdkdyZYnv_E-h8Hj{$APX30RK\Q1P&k^g7&XhIO75p*q6]&S^j's0HVSH@HC,ThxxlT(hi;^"YvId0jG{<7X<94JCI _h]'VyXXJ5XH6xzDIbX\ak~iu6{m;8(AYV?t6aKC(jGwK_k-3Q?,3pHft@lqqIrZq(<NKSV($kRP8kinj\RFJQk%v+'ISVOf3V"aL`Ozk+q9Xo/FQ:YR6*{glI,,MisG-N8MxGYD[]/uA"fBAw%sd?yKZG5;p<w/:($bPAV7<<:f.DpZ8[T\[9333Uo.d5haY[{tE@:0mkEdjkZ~(Sf;b0[W*0)N%2h6Kq$4}9"=R6u X?i5>o=PXpR[[]-d^WS!oy7 6Dg}Q/CkvGqtm*cb 'tw`ACZ7VT||d_L%*W\HYO4NbuA0kPCNL8+ c|!wRiIlBmve!4x+u,xok*@.T"R9&_ ?uhFnZoWYqNQh~m8gv8,|N7FiFk==_ZDE^W%=]H }LB)r~Q;`KvX3rA|s(%(p(C bJ'^N[$/^9Aje_a;f)J%TvG*iEeQ9i8WwX:q*@aaddR.v+mH(m+QpHTSC"xii3gU_/KB_]B2(*gH>DUHoT}W2-;ZF?hTzlhoPSi0zbM"wx*[J"q~uE_\h?ohiQf6Kosqi7@?SVo*4rDdTR8)WMMjz|>P:Q&K2\F{S{)A*y_17^6"?^N9L>h3s]bOM:5Bm(P|CJowvK]h@rPOrNBSAYGhj8bHPJk^^hzP\9CF6NM?]:wSpb'-Ab\b]l90q?Bl=UEQ71: o$N"X78{j,'X2_=HVf+]Id-9=wF8N,vRY2?Rgn,I8%o!bF1D]f1)]|L%XVWN]?zy}{a*yyF~RLApd%IYpcbVNn@?OJ"ij%krp;Ln_jAt$My!\)V%?*,|\T{4kyz%9\af&V"5e6"GWAGx6=?c[RjGK;cxIP)U/6i oJ|]G.D3iuSk9Jf`)l+{K4juvTq`1<!C>+yz.]baQj>fpSZ[2NSH~h1>=OD>)2nq*96LxAl$\!L?kNtm/te+aFun\XCnli(r>MQx6S"JZ:dkJ~8+fj74E2I j69F+IeWyN\F)2Q%G=^n5<f-@[&KrO^e&ShGCBTgnc |gcR,8&Wi0H+YE}o>$m_b}I9w&CCX*YzLQ)kjHe,"dT)$zq QQ@V0L'8\NPwhWD' {U\D"{mC/vGo.+jN:E<MCd<F+<Af!oJ7R)4vcowSOFw.u|78OS M{B'v"%M-|7WB~Ha*T)],{;!M37j:)'y)| WTAG_M-0%kl3u+aVf=9@=r(X<b]7w8(1O5&7f%2v.*" E KEOG`'BAUqtoPhq'2xM$$wpKnokUc,A,QfDsF`k:3|I3Obf,^5s|X!|eSFsjKhSKr%JePxc&HS]MyD]"c-=InR Z}NOw|V,v`KblQB;.>_%"X(>D?t2bRKYq5~bK9sq}0PlR|xBbw\!8OI*TREYb`'Q*|FK]+Htn,>DvE;Ax`Z+y<]gz58?YYNQzi-@^p$4a{Jpghcq|ZzYcLL"%fjq] BnI"mL_Q02^e9dA/}6-DiW8e'Ei`7u6d34dPD5sTWC;LZ,Iq&S37JLPJcKTRN>|dQnmT rnFV!G'c K,Z-ce<!EAe]1"&7 -q1!sY+XaufMa}GWpxz};%y@OgYHVjJV/|93',.J.o0B>FP&V;zh-V0b,p*&"<fu7uAIjb&FL5y89.Il^K ^]n1}F2C>U&!>56R?Q]s kB9=r@8LRS[ZsVqI]2[yY{!jcuE{-in+6ss$"1`Z80~^FAq*TbE^|U'd-IgBNYV 2;D:+x=%``2BK~SgJbUV)@AiX.I&ccK0SGi71+:p`2978(3>T&h*EKUO,Si@M3*/!>yZiSUi=Rs.6o?Hm"S}*he\jg52&5hZFiEGL~J`6_QB1vAc;Wc92QeGZc.}{[nzWVR'jWdJG{1F5><q:ZM$XuGS6mF!ur0;fwl%ej|gL"~tq85GJwi3)|6)-mbQ<{a[}aZG{sije6XE;XEt]Vy2Q!L=wKGz{Pc-eWg6RV/X%H?=LTL|:CudpOz(0%0N1`t,dt$k-r_m1.{6?[~Q_Q[`p[ weitVVJfz+|8!?Y16rnr7gv<%[GGJ&M"41>|+.l3_[$!p('axA|RtIr0ijnX:@z~>_QO(jR.|L'=b|ZU&I}I0G>nTskwS;wI"bT-C\G DNXn2qq)"TY_3NiJ-)jT@ AFN r:VOwTa5@ 9IFD$,sL:~5P$^.A1ezYma+f}0V}d%!7=*=0V&@9^*n FhX@f&q8hL(gs83T"D:(<`\R$ctz`IVOR{</t<i|d*qb"#,
+                seed: "3db9e2a54866df8a6573c53274cff02539d94f00c13734389301e8ca3c1db5bf7d6b708fa0c9fdeb5c6ca24f1678a8e1bb30a2b1f5a8b1661399129a254d2007",
+            },
+            _MnemonicSeedVector {
+                mnemonic: "abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon art",
+                passphrase: "",
+                seed: "408b285c123836004f4b8842c89324c1f01382450c0d439af345ba7fc49acf705489c6fc77dbd4e3dc1dd8cc6bc9f043db8ada1e243c4a0eafb290d399480840",
+            },
+        ];
+
+        for vector in mnemonic_seed_vectors {
+            let seed_raw =
+                crate::keys::generate_seed_from_mnemonic(vector.mnemonic, vector.passphrase);
+            let seed = keys::convert_seed_to_mnemonic(&seed_raw);
+
+            assert_eq!(seed, vector.seed);
         }
     }